fix macOS bundle working directory
